Update to parquet-format 2.11.0 (#7570)
# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax. For example
`Closes #123` indicates that this PR will close issue #123.
-->

Closes #7567.

# Rationale for this change

Update to latest parquet format.
<!--
Why are you proposing this change? If this is already explained clearly
in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand
your changes and offer better suggestions for fixes.
-->

# What changes are included in this PR?

Re-generate format.rs and make any needed changes to allow compilation.
No support for new logical types is added.
<!--
There is no need to duplicate the description in the issue here but it
is sometimes worth providing a summary of the individual changes in this
PR.
-->

# Are there any user-facing changes?
yes

<!--
If there are user-facing changes then we may require documentation to be
updated before approving the PR.
-->

<!---
If there are any breaking changes to public APIs, please call them out.
-->