Add `TimeMilli` and `TimeMicro` fields and conversions for the record API  (#7544)
# Which issue does this PR close?

- Closes #7543

# Rationale for this change
 
I recently wrote a PR for https://github.com/apache/arrow-rs/issues/7510
that was wrong. So this is rectifying that.

# What changes are included in this PR?

I had used TimestampMillis for a time value that does not have a date,
which is wrong.

I've added two new Fields for the missing concept of "time without
date", and fixed the conversion to use them, and added the required
trait implementations

# Are there any user-facing changes?


<!--
If there are user-facing changes then we may require documentation to be
updated before approving the PR.
-->

There are new `Field` variants, so if people are matching `Field`,
this'll cause a compile error when they upgrade

---------

Co-authored-by: Ed Seidl <etseidl@live.com>