feat: add constructor to efficiently upgrade dict key type to remaining builders (#7689)
# Which issue does this PR close?

Closes https://github.com/apache/arrow-rs/issues/7654

# Rationale for this change

In https://github.com/apache/arrow-rs/pull/7611 we added a constructor
to the GenericByteDictionaryBuilder to create a new builder instance
from an existing one, but with a modified type for the key. It would be
good if other dictionary builders had similar capability.

# What changes are included in this PR?

This followup PR adds this capability to remaining builder types

# Are there any user-facing changes?

No