Implement Type and Array for Decimal32 and Decimal64 (#7098)
# Which issue does this PR close?

Part of addressing #6661 but does not close it.

# What changes are included in this PR?

This change implements `Decimal32Array`, `Decimal32Type`,
`Decimal64Array` and `Decimal64Type` in a relatively minimal fashion. It
doesn't implement the full range of casts or support for Parquet or CSV
reader. These will come in the next change.

# Are there any user-facing changes?

Adds several public types. Adds new entries for the `DataType` enum,
which is a breaking change.