Perf: Make sort string view fast(1.5X ~ 3X faster) (#7792)
# Which issue does this PR close?

This is a follow-up for https://github.com/apache/arrow-rs/pull/7748

In theory we can custom string view compare, and make it crazy faster.

- Closes [#7790](https://github.com/apache/arrow-rs/issues/7790)

# Rationale for this change

In theory we can custom string view compare, and make it crazy faster.

# What changes are included in this PR?

In theory we can custom string view compare, and make it crazy faster.

# Are these changes tested?

Yes

# Are there any user-facing changes?

No