[VARIANT] Add support for the json_to_variant API (#7783)
# Which issue does this PR close?

- Closes #7425.

# Rationale for this change

Explained in the issue.

# What changes are included in this PR?

This PR includes a `json_to_variant` API to parse JSON strings as
Variants. `json_to_variant` takes as argument the input JSON string and
an object `builder` of type `VariantBuilder` and builds the variant
using `builder`. The resulting variant can be extracted using
`builder.finish()` which consumes the builder and returns the Variant
buffers.

# Are these changes tested?

Unit Tests, and an example file.

# Are there any user-facing changes?

Yes, the PR introduces the `json_to_variant` API.

---------

Co-authored-by: Andrew Lamb <andrew@nerdnetworks.org>
Co-authored-by: Ryan Johnson <scovich@users.noreply.github.com>