Add own_values() and own_values_offsets() methods
The problem with sliced arrays is that its values() method
returns the original values slice due to zero-copy implementation
of slice(). The same issue occurs when trying to observe
sliced array offsets: the offsets do not start from 0.

The change adds methods for accessing own values in O(1) and
own offsets (starting from 0).