[Variant] Introduce `parquet-variant-json` crate (#7862)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Closes https://github.com/apache/arrow-rs/issues/7800
- Closes https://github.com/apache/arrow-rs/pull/7845
- Closes https://github.com/apache/arrow-rs/issues/7775

# Rationale for this change

Now that we have the basic json conversion functionality complete thanks
to @harshmotw-db ❤️ ❤️ ❤️ in
- https://github.com/apache/arrow-rs/pull/7783

I would like to move all the json related code into its own crate to
keep the functionality clean and clear

# What changes are included in this PR?

1. Move to_json and from_json into a new `parquet-variant-json` crate
where we can continue to iterate
2. Added new crate to CI jobs


# Are these changes tested?

CI
# Are there any user-facing changes?

If there are user-facing changes then we may require documentation to be
updated before approving the PR.

If there are any breaking changes to public APIs, please call them out.