[Variant] Define basic convenience methods for variant pathing (#7894)
# Which issue does this PR close?

Part of
* https://github.com/apache/arrow-rs/issues/6736

# Rationale for this change

An expected use case for variant pathing (e.g. a future `variant_get`)
would be to request a field from a variant value that is expected to be
an object, or an element from a variant value that is expected to be an
array. Those methods are currently missing.

# What changes are included in this PR?

Define `Variant::get_object_field` and `Variant::get_array_element`
methods that do what they say (returning `None` if anything mismatches).

# Are these changes tested?

New doc tests for the methods.

# Are there any user-facing changes?

New public methods on `Variant` enum.