Implement arrow-avro Reader and ReaderBuilder (#7834)
# Which issue does this PR close?

Part of https://github.com/apache/arrow-rs/issues/4886

# Rationale for this change

This PR refactors the Avro reader's public API to provide an ergonomic
and idiomatic experience.
The new API provides a high-level and fluent `Reader` and
`ReaderBuilder`, making it more consistent with other readers in the
ecosystem.

# What changes are included in this PR?

The core of this PR is the introduction of a struct-based API design for
the arrow-avro reader.
- **`Reader<R: BufRead>`**: The main addition is a high-level struct
that implements the standard and traits. This encapsulates the entire
reading process, allowing users to simply iterate over the reader.
- **`ReaderBuilder`**: This provides a fluent interface for configuring
and constructing a `Reader`. It simplifies the setup for options such
as: `batch_size`, `strict_mode`, `use_utf8view`, `with_schema`.
- **Streaming `Decoder`**: For more advanced, asynchronous use cases, a
streaming decoder has been introduced. It's designed to be used in
streaming contexts where Avro data arrives in chunks.

# Are these changes tested?

Yes. The existing tests, such as `test_alltypes` and
`test_utf8view_support` have been refactored to use the new API. This
refactor validates the new abstractions. Additionally, new tests have
been added to cover the streaming functionality, ensuring it works
correctly.

# Are there any user-facing changes?

N/A

# Follow-Up PRs

1. Add Remaining `arrow-avro` Compression types.
2. Complete `strict_mode` on `arrow-avro` ReaderBuilder For Impala Avro
Support
3. Add Remaining `arrow-avro` Reader Integration Tests.
4. Complete `arrow-avro` Empty Record Decoding.
5. Implement `arrow-avro` Decoder Schema Resolution. (This one could be
broken up depending on size and the `arrow-avro` Reader should be ready
to go public upon completion)

---------

Co-authored-by: Andrew Lamb <andrew@nerdnetworks.org>
Co-authored-by: Ryan Johnson <scovich@users.noreply.github.com>