[Variant] Use simdutf8 for UTF-8 validation (#7908)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Closes #7902.

# Rationale for this change


# What changes are included in this PR?

- add `simdutf8` dependency for parquet-variant
- add a fn `extract_and_validate_utf8_slice` 

# Are these changes tested?

We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?

# Are there any user-facing changes?

If there are user-facing changes then we may require documentation to be
updated before approving the PR.

If there are any breaking changes to public APIs, please call them out.

---------

Signed-off-by: codephage2020 <tingwangyan2020@163.com>
Co-authored-by: Andrew Lamb <andrew@nerdnetworks.org>