[Variant] Impl `PartialEq` for VariantObject (#7943)
# Rationale for this change

- Closes https://github.com/apache/arrow-rs/issues/7948

This PR introduces a custom implementation of `PartialEq` for variant
objects.

According to the spec, field values are not required to be in the same
order as the field IDs, to enable flexibility when constructing Variant
values.

Instead of comparing the raw bytes of 2 variant objects, this
implementation recursively checks whether the field values are equal --
regardless of their order