Add additional integration tests to arrow-avro (#7974)
# Which issue does this PR close?

Part of https://github.com/apache/arrow-rs/issues/4886

Completes the breaking down/porting of the changes in
https://github.com/apache/arrow-rs/pull/6965. This PR will be closed
upon merge of this PR.

# Rationale for this change

This change brings over the remaining integration tests present in the
original PR, which validate the reader logic against the files from
`testing/data/avro`. PRs containing this logic have already been merged
(but are not yet released) which these tests now validate.

# What changes are included in this PR?

The following files are now read in:

- alltypes_dictionary.avro
- alltypes_nulls_plain.avro
- binary.avro
- dict-page-offset-zero.avro
- avro/list_columns.avro
- nested_lists.snappy.avro
- single_nan.avro
- datapage_v2.snappy.avro
- nested_records.avro
- repeated_no_annotation.avro

# Are these changes tested?

This PR consists of integration tests validating code merged recently
into this crate. No changes in functionality are included.

# Are there any user-facing changes?

N/A