arrow-schema: Remove dict_id from being required equal for merging (#7968)
# Which issue does this PR close?

Closes https://github.com/apache/arrow-rs/issues/6356

# Rationale for this change

Now that https://github.com/apache/arrow-rs/pull/7940 is merged, nothing
useful can be done with the `dict_id` field, therefore, it is now safe
to be removed from this requirement.

This was also split out from:
https://github.com/apache/arrow-rs/pull/7467

# What changes are included in this PR?

No longer require the `dict_id` fields of two `Field`s of schemas being
merged to be equal, as at this point the `dict_id` is only an IPC
concern, and the fact that it is still in the struct definition is just
legacy, marked for removal, we're just going through the proper
procedure of deprecating and replacing the APIs that use it.

# Are these changes tested?

Tests passing.

# Are there any user-facing changes?

No API changes, just a behavior change, that was to be expected and
desired due to the deprecations around the `dict_id` field.

@alamb @adriangb @tustvold