Add decimal32 and decimal64 support to Parquet, JSON and CSV readers and writers (#7841)
# Which issue does this PR close?

- Finishes remaining work and closes #6661.

# What changes are included in this PR?

This change adds `decimal32` and `decimal64` support to Parquet, JSON
and CSV readers and writers. It does not change the current default
behavior of the Parquet reader which (in the absence of a specification
that says otherwise) will still translate the INT32 physical type with a
logical DECIMAL type into a `decimal128` instead of a `decimal32`.

# Are these changes tested?

Yes.

# Are there any user-facing changes?

The `decimal32` and `decimal64` types are now supported in Parquet, JSON
and CSV readers and writers.

---------

Co-authored-by: Andrew Lamb <andrew@nerdnetworks.org>
Co-authored-by: Matthijs Brobbel <m1brobbel@gmail.com>