[Parquet] Allow writing compatible DictionaryArrays to parquet writer (#8005)
# Which issue does this PR close?

- closes https://github.com/apache/arrow-rs/issues/8004

# Rationale for this change

I'm opening this PR to explore the idea that for a given column, the
`ArrowWriter` should compatible consider either native arrays or
dictionary arrays with the same value.

e.g. if I write two record batches, and in the first batch column `a` is
type `Dictionary<_, Utf8>`, I should be able to append a second batch to
my writer where column `a` is type `Utf8` (the native array).

# What changes are included in this PR?

This PR relaxes the added in
https://github.com/apache/arrow-rs/pull/5341 when creating
`LevelsInfoBuilder` to consider the types compatible using the logic
explained above.

# Are these changes tested?

There is a basic unit test. I'm happy to add more if this change is
something acceptable.

# Are there any user-facing changes?
No