implement `cast_to_variant` kernel to cast native types to `VariantArray` (#8044)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Closes https://github.com/apache/arrow-rs/issues/8043


# Rationale for this change


As @Samyak2 suggested on
https://github.com/apache/arrow-rs/pull/8021/files#r2249926579, having
the ability to convert *FROM* a typed value to a VariantArray will be
important

For example, in SQL it could be used to cast columns to variant like in
`some_column::variant`


# What changes are included in this PR?

1. Add `cast_to_variant` kernel to cast native types to `VariantArray`
2. Tests


# Are these changes tested?
yes


# Are there any user-facing changes?

New kernel