Minor: Consolidate int96 stats roundtrip test (#8034)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Closes #NNN.

# Rationale for this change

I was slightly annoyed that the int96 stats roundtrip test was not part
of the other arrow-reader tests

You have to run it separately:
```shell
cargo test --test int96_stats_roundtrip
```

Let's just move it to the arrow-reader tests, so it is run with the
other tests so
1. It is easier to find
2. It takes slightly less time to build and test the parquet crate



```shell
cargo test --test arrow_reader
```


# What changes are included in this PR?
1. Move the tests

# Are these changes tested?

Yes, by CI

# Are there any user-facing changes?
No