fix: Validate metadata len in IPC reader  (#8097)
# Which issue does this PR close?

No issue filed.

# Rationale for this change

We allocate memory based on metadata length - If an untrusted client
writes a meta len of < 0 then we'll allocate a large number of bytes due
to sign extension and likely panic.

# What changes are included in this PR?

- Update StreamReader in both places it reads metadata length for < 0
which is at the start of the stream to read the schema, and in the
middle of the stream between each message.

# Are these changes tested?

Yes, tests for both reads are added

# Are there any user-facing changes?

No

---------

Co-authored-by: Jake Dern <jakedern@microsoft.com>
Co-authored-by: Liang-Chi Hsieh <viirya@gmail.com>