[thrift-remodel] Redo thrift enums and unions (#8072)
# Which issue does this PR close?
**Note: this targets a feature branch, not main**

- Part of #5854.
- Related to #6129

# Rationale for this change

Next step.

# What changes are included in this PR?

Attempt to make use of macros originally developed by @jhorstmann to
implement some thrift enums and unions in `basic.rs`. Also adds yet
another spin on a thrift decoder based heavily on
`TCompactSliceInputProtocol`. The current approach is to use this new
decoder with `TryFrom` impls on the data structures in question.

This PR does not yet complete the process far enough to parse a parquet
footer, but I'm putting it out here to get early feedback on the design.

Some structures already deviate enough from their thrift counterparts
that the macro based parsing is not practical. Likewise, the macro
approach doesn't work for structs that need lifetime annotations
(necessary for reading binary fields as slices).

# Are these changes tested?

Not yet.

# Are there any user-facing changes?

Yes.