[Variant] Add human-readable impl Debug for Variant (#8140)
# Which issue does this PR close?

- Related to https://github.com/apache/arrow-rs/issues/8136

# Rationale for this change

Unit tests need a way to verify two `Variant` are logically equivalent,
and `Debug` is a good way to achieve that without wading into the
complexities of a proper `PartialEq` implementation that would become
part of the public API.

More generally, byte slices are not very easy for humans to interpret,
so it makes sense for `Debug` to do something nicer.

# What changes are included in this PR?

Manually `impl Debug for Variant`, maintaining a traditional look but
with nicer handling of `Variant::Binary`, `Variant::Object` and
`Variant::List` subtypes. The latter two use fallible iteration to avoid
potential panics, since `Debug` is likely to be used when formatting
error messages.

# Are these changes tested?

New unit test

# Are there any user-facing changes?

The debug formatting of `Variant` has changed.