[thrift-remodel] Complete decoding of `FileMetaData` and `RowGroupMetaData` (#8111)
# Which issue does this PR close?
**Note: this targets a feature branch, not main**

- Part of #5854.

# Rationale for this change

# What changes are included in this PR?
This PR completes reading of the `FileMetaData` and `RowGroupMetaData`
pieces of the `ParquetMetaData`. Column indexes and encryption will be
follow-on work.

This replaces the macro for generating structs with a more general one
that can take visibility and lifetime specifiers. Also (temporarily)
adds a new function `ParquetMetaDataReader::decode_file_metadata` which
should be a drop-in replacement for
`ParquetMetaDataReader::decode_metadata`.

Still todo:

1. Add some tests that verify this produces the same output as
`ParquetMetaDataReader::decode_metadata`
2. Read column indexes with new decoder
3. Read page headers with new decoder
4. Integrate with @alamb's push decoder work #8080

# Are these changes tested?

Not yet

# Are there any user-facing changes?

Yes