[Parquet] Add tests for IO/CPU access in parquet reader (#7971)
# Which issue does this PR close?

- Part of https://github.com/apache/arrow-rs/issues/8000

- Related to https://github.com/apache/arrow-rs/pull/7850

# Rationale for this change

There is quite a bit of code in the current Parquet sync and async
readers related to IO patterns that I do not think is not covered by
existing tests. As I refactor the guts of the readers into the
PushDecoder, I would like to ensure we don't introduce regressions in
existing functionality.

I would like to add tests that cover the IO patterns of the Parquet
Reader so I don't break it

# What changes are included in this PR?

Add tests which
1. Creates a temporary parquet file with a known row group structure
2. Reads data from that file using the Arrow Parquet Reader, recording
the IO operations
3. Asserts the expected IO patterns based on the read operations in a
human understandable behavior

This is done for both the sync and async readers.

I am sorry this is such a massive PR, but it is entirely tests and I
think it is quite important. I could break the sync or async tests into
their own PR, but this seems uncessary

# Are these changes tested?

Yes, indeed the entire PR is only tests


# Are there any user-facing changes?