Docs: Clarify that Array::value does not check for nulls (#8065)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Related to https://github.com/apache/arrow-rs/pull/8021

# Rationale for this change

As part of the review in https://github.com/apache/arrow-rs/pull/8021,
@scovich and I were discussing how `VariantArray::value` should behave
in the presence of nulls:
https://github.com/apache/arrow-rs/pull/8021#discussion_r2256914173

> Suggest to make this return Option<Variant> so callers don't have to
check for null themselves.

I realized it might not be 100% clear that the existing convention in
this crate was that `value()` methods did not check for nulls / return
`Option`. I think we should document it better

# What changes are included in this PR?

Explicitly document that `value` methods do not check for nulls and
explain what happens when they are used on null values

# Are these changes tested?

Yes, by CI

# Are there any user-facing changes?

Additional documentation. No behavior changes

---------

Co-authored-by: Kyle Barron <kylebarron2@gmail.com>