[Variant] Rename `batch_json_string_to_variant` and `batch_variant_to_json_string` (#8161)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Closes #8144.

# Rationale for this change

# What changes are included in this PR?

Use extension traits to wrap the json variant conversion function, and
rename batch function to a more common name.

# Are these changes tested?

Yes

# Are there any user-facing changes?

The APIs of parquet-variant-json are changed

---------

Co-authored-by: Andrew Lamb <andrew@nerdnetworks.org>