Implement cast and other operations on decimal32 and decimal64 (#7815)
# Which issue does this PR close?

Part of addressing #6661 but does not close it; there is at least one
more PR for CSV/Parquet/JSON support.

# What changes are included in this PR?

This change adds cast operations for the recently-added `decimal32` and
`decimal64` types. It also adds tests to verify that sort and comparison
work correctly, and includes these types in benchmarks.

# Are these changes tested?

Yes.

# Are there any user-facing changes?

Casting to and from `decimal32` and `decimal64` is now supported.