[thrift-remodel] Decoding of page indexes (#8160)
# Which issue does this PR close?

**Note: this targets a feature branch, not main**

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Part of #5854.

# Rationale for this change

Speed

# What changes are included in this PR?

Still a work in progress, but begins the process of converting page
index parsing to the new thrift decoder.

# Are these changes tested?

This PR actually uses the new decoder when parsing the page indexes
using the existing machinery. As such all tests involving the page
indexes should apply to this code.

# Are there any user-facing changes?

Yes