arrow-row: Document dictionary handling (#8168)
# Which issue does this PR close?

- related to https://github.com/apache/arrow-rs/pull/7627
- Related to https://github.com/apache/arrow-rs/issues/4811

# Rationale for this change

It was not clear to me what the expected behavior for round trip through
row converter was for DictionaryArrays, so let's document what @tustvold
says here:
https://github.com/apache/arrow-rs/pull/8067#issuecomment-3160849076

> I think the issue is that Datafusion is not handling the fact that row
encoding "hydrates" dictionaries. It should be updated to understand
that List<Dictionary<...>> will be converted to List<...>, much like it
already handles this for the non-nested case. Converting back to a
dictionary is expensive, and likely pointless, not to mention a breaking
change.


# What changes are included in this PR?

Document expected behavior with english comments and doc test

# Are these changes tested?

Yes (doctests)

# Are there any user-facing changes?

More docs, no behavior change