Implement `ArrayBuilder` for `UnionBuilder` (#8169)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Closes #8033.

# What changes are included in this PR?

* Make `FieldDataValues: Send + Sync`
* Derive `Default` for `UnionBuilder`
* Implement `build_cloned` for `UnionBuilder`

# Are these changes tested?

Yes.

# Are there any user-facing changes?

These changes should be backwards compatible and don't change the
existing public API.

Co-authored-by: Andrew Lamb <andrew@nerdnetworks.org>