[Variant] VariantArrayBuilder tracks only offsets (#8193)
# Which issue does this PR close?

- Closes https://github.com/apache/arrow-rs/issues/8192

# Rationale for this change

Tracking lengths is redundant because that's just the difference between
adjacent offsets. It becomes even easier if we eventually start adding
the views eagerly instead of storing up offsets, because we only need to
remember the difference between starting and final offset.

# What changes are included in this PR?

Update the algorithm that creates the binary view to compute the
offsets.

# Are these changes tested?

Existing tests cover this change.

# Are there any user-facing changes?

No.